Something new, adding some typedef structs
