+++ include/Alert.h
@@ -1,0 +1,21 @@
+#ifndef A_ALERT_h
+#define A_ALERT_h
+
+#include "enum.h"
+
+
+/***********************
+ * GENERAL ALERT TYPES *
+ ***********************/
+enum alertType {
+    ALERT_DEFAULT = 0,
+    ALERT_PULSE_SLOW,
+    ALERT_PULSE_FAST,
+    ALERT_ON,
+    ALERT_OFF,
+    ALERT_NONE
+};
+
+
+
+#endif
