+++ include/Logging.h
@@ -10,5 +10,17 @@
 #include <ControlIO.h>
 #include <enum.h>
 
+/******************
+ * LOGGING STATES *
+ ******************/
+enum loggingState {
+    LOG_NONE = 0,
+    LOG_ENABLED,        // Logging Enabled
+    LOG_DISABLED,       // Logging Disabled
+    LOG_ERR             // Logging Error
+};
+
+
+
 
 #endif
